Fix incorrect type for get_min/max_length

+++ src/quicktype-core/language/CPlusPlus.ts
@@ -299,8 +299,8 @@ type ConstraintMember = {
 const constraintMembers: ConstraintMember[] = [
     { name: MemberNames.MinValue, getter: MemberNames.GetMinValue, setter: MemberNames.SetMinValue, cppType: "int" },
     { name: MemberNames.MaxValue, getter: MemberNames.GetMaxValue, setter: MemberNames.SetMaxValue, cppType: "int" },
-    { name: MemberNames.MinLength, getter: MemberNames.GetMinLength, setter: MemberNames.SetMinLength, cppType: "int" },
-    { name: MemberNames.MaxLength, getter: MemberNames.GetMaxLength, setter: MemberNames.SetMaxLength, cppType: "int" },
+    { name: MemberNames.MinLength, getter: MemberNames.GetMinLength, setter: MemberNames.SetMinLength, cppType: "size_t" },
+    { name: MemberNames.MaxLength, getter: MemberNames.GetMaxLength, setter: MemberNames.SetMaxLength, cppType: "size_t" },
     {
         name: MemberNames.Pattern,
         getter: MemberNames.GetPattern,
